Introduce a duplex wait indicator
This one is for asynchronous waiting where progress _can_ continue on a
wake / remote progress but it does not necessitate it, i.e. also waiting
on further data to bump on their own side of the stream (from freed
buffers, or more produced data). That allows a more effective waiting
pattern where both modes of progress are awaited.