Split crate into lib / bin
The binary for server is *not* normative and brings in a host of
dependencies, optional but still marked in the cargo tree, that we do
not actually need for functioning. This also avoids some feature
explosion as the `bin` feature is redundant.

You will be encouraged to spawn a server in some host process and
provide the socket via non-CLOEXEC instead. That pattern works well with
other types of shared memory such as memfd-secret that must be
restricted in access. A server binary unduly presupposes that a separate
process for handling the host functioning is recommended.

As a consequence, we shall turn the basic control loop for the server
into a library convenience function provided with io-uring.